skills/pptx/scripts/requirements.txt
python-pptx>=0.6.23
Pillow>=10.0.0
lxml>=5.0.0
defusedxml>=0.7.1
